// tests/protoc-gen-ts-grpc-gateway-cli/integration-everything-simple/caseUseProtoNames/proto/bodyjson/bodyjson.ts
export interface WellKnownTypesHolder {
  payloadCheck: string;
  any?: Any | undefined;
  empty?: Empty | undefined;
  struct?: { [key: string]: any } | undefined;
  value?: any | undefined;
  listValue?: Array<any> | undefined;
  timestamp?: Date | undefined;
  duration?: Duration | undefined;
  fieldMask?: string[] | undefined;
  boolValue?: boolean | undefined;
  int32Value?: number | undefined;
  uint32Value?: number | undefined;
  int64Value?: number | undefined;
  uint64Value?: number | undefined;
  stringValue?: string | undefined;
  bytesValue?: Uint8Array | undefined;
  doubleValue?: number | undefined;
  floatValue?: number | undefined;
}
// ...
export const WellKnownTypesHolder: MessageFns<WellKnownTypesHolder> = {
  fromJSON(object: any): WellKnownTypesHolder {
    return {
      payloadCheck: isSet(object.payload_check)
        ? globalThis.String(object.payload_check)
        : "",
      any: isSet(object.any) ? Any.fromJSON(object.any) : undefined,
      empty: isSet(object.empty) ? Empty.fromJSON(object.empty) : undefined,
      struct: isObject(object.struct) ? object.struct : undefined,
      value: isSet(object?.value) ? object.value : undefined,
      listValue: globalThis.Array.isArray(object.list_value)
        ? [...object.list_value]
        : undefined,
      timestamp: isSet(object.timestamp)
        ? fromJsonTimestamp(object.timestamp)
        : undefined,
      duration: isSet(object.duration)
        ? Duration.fromJSON(object.duration)
        : undefined,
      fieldMask: isSet(object.field_mask)
        ? FieldMask.unwrap(FieldMask.fromJSON(object.field_mask))
        : undefined,
      boolValue: isSet(object.bool_value)
        ? Boolean(object.bool_value)
        : undefined,
      int32Value: isSet(object.int32_value)
        ? Number(object.int32_value)
        : undefined,
      uint32Value: isSet(object.uint32_value)
        ? Number(object.uint32_value)
        : undefined,
      int64Value: isSet(object.int64_value)
        ? Number(object.int64_value)
        : undefined,
      uint64Value: isSet(object.uint64_value)
        ? Number(object.uint64_value)
        : undefined,
      stringValue: isSet(object.string_value)
        ? String(object.string_value)
        : undefined,
      bytesValue: isSet(object.bytes_value)
        ? new Uint8Array(object.bytes_value)
        : undefined,
      doubleValue: isSet(object.double_value)
        ? Number(object.double_value)
        : undefined,
      floatValue: isSet(object.float_value)
        ? Number(object.float_value)
        : undefined,
    };
  },
// ...
};
// ...
export interface BodyJSONServiceClient<CallOptionsExt = {}> {
  postEnumBody(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  postStringBody(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  postRepeatedMessageBody(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  postRepeatedEnumBody(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  postRepeatedStringBody(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  postTimestampBody(
    request: DeepPartial<WellKnownTypesHolder>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<WellKnownTypesHolder>;
  postFieldMaskBody(
    request: DeepPartial<WellKnownTypesHolder>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<WellKnownTypesHolder>;
  postStructBody(
    request: DeepPartial<WellKnownTypesHolder>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<WellKnownTypesHolder>;
  postValueBody(
    request: DeepPartial<WellKnownTypesHolder>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<WellKnownTypesHolder>;
  postListValueBody(
    request: DeepPartial<WellKnownTypesHolder>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<WellKnownTypesHolder>;
  postWrapperBody(
    request: DeepPartial<WellKnownTypesHolder>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<WellKnownTypesHolder>;
}

type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;