// tests/protoc-gen-ts-grpc-gateway-cli/integration-everything-simple/caseUseProtoNames/proto/examplepb/a_bit_of_everything.ts
export enum NumericEnum {
  /** ZERO - ZERO means 0 */
  ZERO = 0,
  /** ONE - ONE means 1 */
  ONE = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface ABitOfEverything {
  singleNested: ABitOfEverything_Nested | undefined;
  uuid: string;
  nested: ABitOfEverything_Nested[];
  floatValue: number;
  doubleValue: number;
  int64Value: number;
  uint64Value: number;
  int32Value: number;
  fixed64Value: number;
  fixed32Value: number;
  boolValue: boolean;
  stringValue: string;
  bytesValue: Uint8Array;
  uint32Value: number;
  enumValue: NumericEnum;
  pathEnumValue: PathEnum;
  nestedPathEnumValue: MessagePathEnum_NestedPathEnum;
  sfixed32Value: number;
  sfixed64Value: number;
  sint32Value: number;
  sint64Value: number;
  repeatedStringValue: string[];
  oneofEmpty?: Empty | undefined;
  oneofString?: string | undefined;
  mapValue: { [key: string]: NumericEnum };
  mappedStringValue: { [key: string]: string };
  mappedNestedValue: { [key: string]: ABitOfEverything_Nested };
  nonConventionalNameValue: string;
  timestampValue: Date | undefined;
  /** repeated enum value. it is comma-separated in query */
  repeatedEnumValue: NumericEnum[];
  /** repeated numeric enum comment (This comment is overridden by the field annotation) */
  repeatedEnumAnnotation: NumericEnum[];
  /** numeric enum comment (This comment is overridden by the field annotation) */
  enumValueAnnotation: NumericEnum;
  /** repeated string comment (This comment is overridden by the field annotation) */
  repeatedStringAnnotation: string[];
  /** repeated nested object comment (This comment is overridden by the field annotation) */
  repeatedNestedAnnotation: ABitOfEverything_Nested[];
  /** nested object comments (This comment is overridden by the field annotation) */
  nestedAnnotation: ABitOfEverything_Nested | undefined;
  int64OverrideType: number;
  /** mark a field as required in Open API definition */
  requiredStringViaFieldBehaviorAnnotation: string;
  /** mark a field as readonly in Open API definition */
  outputOnlyStringViaFieldBehaviorAnnotation: string;
  optionalStringValue?: string | undefined;
  /** Test openapiv2 generation of repeated fields */
  productId: string[];
  /** Test openapiv2 generation of required fields with annotation and jsonschema to reproduce */
  optionalStringField: string;
  requiredStringField1: string;
  requiredStringField2: string;
  /** Test openapiv2 handling of required json_name fields */
  requiredFieldBehaviorJsonName: string;
  requiredFieldSchemaJsonName: string;
  /** Trailing only */
  trailingOnly: string;
  /** Trailing only dot. */
  trailingOnlyDot: string;
  /** Leading both */
  trailingBoth: string;
  /**
   * Leading multiline
   *
   * This is an example of a multi-line comment.
   */
  trailingMultiline: string;
  /** Specify a custom format of repeated field items */
  uuids: string[];
}

/** Nested is nested type. */
export interface ABitOfEverything_Nested {
  /** name is nested field. */
  name: string;
  amount: number;
  /** DeepEnum comment. */
  ok: ABitOfEverything_Nested_DeepEnum;
}

/** DeepEnum is one or zero. */
export enum ABitOfEverything_Nested_DeepEnum {
  /** FALSE - FALSE is false. */
  FALSE = 0,
  /** TRUE - TRUE is true. */
  TRUE = 1,
  UNRECOGNIZED = -1,
}
// ...
export const ABitOfEverything: MessageFns<ABitOfEverything> = {
  fromJSON(object: any): ABitOfEverything {
    return {
      singleNested: isSet(object.single_nested)
        ? ABitOfEverything_Nested.fromJSON(object.single_nested)
        : undefined,
      uuid: isSet(object.uuid) ? globalThis.String(object.uuid) : "",
      nested: globalThis.Array.isArray(object?.nested)
        ? object.nested.map((e: any) => ABitOfEverything_Nested.fromJSON(e))
        : [],
      floatValue: isSet(object.float_value)
        ? globalThis.Number(object.float_value)
        : 0,
      doubleValue: isSet(object.double_value)
        ? globalThis.Number(object.double_value)
        : 0,
      int64Value: isSet(object.int64_value)
        ? globalThis.Number(object.int64_value)
        : 0,
      uint64Value: isSet(object.uint64_value)
        ? globalThis.Number(object.uint64_value)
        : 0,
      int32Value: isSet(object.int32_value)
        ? globalThis.Number(object.int32_value)
        : 0,
      fixed64Value: isSet(object.fixed64_value)
        ? globalThis.Number(object.fixed64_value)
        : 0,
      fixed32Value: isSet(object.fixed32_value)
        ? globalThis.Number(object.fixed32_value)
        : 0,
      boolValue: isSet(object.bool_value)
        ? globalThis.Boolean(object.bool_value)
        : false,
      stringValue: isSet(object.string_value)
        ? globalThis.String(object.string_value)
        : "",
      bytesValue: isSet(object.bytes_value)
        ? bytesFromBase64(object.bytes_value)
        : new Uint8Array(0),
      uint32Value: isSet(object.uint32_value)
        ? globalThis.Number(object.uint32_value)
        : 0,
      enumValue: isSet(object.enum_value)
        ? numericEnumFromJSON(object.enum_value)
        : 0,
      pathEnumValue: isSet(object.path_enum_value)
        ? pathEnumFromJSON(object.path_enum_value)
        : 0,
      nestedPathEnumValue: isSet(object.nested_path_enum_value)
        ? messagePathEnum_NestedPathEnumFromJSON(object.nested_path_enum_value)
        : 0,
      sfixed32Value: isSet(object.sfixed32_value)
        ? globalThis.Number(object.sfixed32_value)
        : 0,
      sfixed64Value: isSet(object.sfixed64_value)
        ? globalThis.Number(object.sfixed64_value)
        : 0,
      sint32Value: isSet(object.sint32_value)
        ? globalThis.Number(object.sint32_value)
        : 0,
      sint64Value: isSet(object.sint64_value)
        ? globalThis.Number(object.sint64_value)
        : 0,
      repeatedStringValue: globalThis.Array.isArray(
        object?.repeated_string_value,
      )
        ? object.repeated_string_value.map((e: any) => globalThis.String(e))
        : [],
      oneofEmpty: isSet(object.oneof_empty)
        ? Empty.fromJSON(object.oneof_empty)
        : undefined,
      oneofString: isSet(object.oneof_string)
        ? globalThis.String(object.oneof_string)
        : undefined,
      mapValue: isObject(object.map_value)
        ? Object.entries(object.map_value).reduce<{
            [key: string]: NumericEnum;
          }>((acc, [key, value]) => {
            acc[key] = numericEnumFromJSON(value);
            return acc;
          }, {})
        : {},
// ...
      repeatedEnumValue: globalThis.Array.isArray(object?.repeated_enum_value)
        ? object.repeated_enum_value.map((e: any) => numericEnumFromJSON(e))
        : [],
      repeatedEnumAnnotation: globalThis.Array.isArray(
        object?.repeated_enum_annotation,
      )
        ? object.repeated_enum_annotation.map((e: any) =>
            numericEnumFromJSON(e),
          )
        : [],
      enumValueAnnotation: isSet(object.enum_value_annotation)
        ? numericEnumFromJSON(object.enum_value_annotation)
        : 0,
      repeatedStringAnnotation: globalThis.Array.isArray(
        object?.repeated_string_annotation,
      )
        ? object.repeated_string_annotation.map((e: any) =>
            globalThis.String(e),
          )
        : [],
      repeatedNestedAnnotation: globalThis.Array.isArray(
        object?.repeated_nested_annotation,
      )
        ? object.repeated_nested_annotation.map((e: any) =>
            ABitOfEverything_Nested.fromJSON(e),
          )
        : [],
      nestedAnnotation: isSet(object.nested_annotation)
        ? ABitOfEverything_Nested.fromJSON(object.nested_annotation)
        : undefined,
      int64OverrideType: isSet(object.int64_override_type)
        ? globalThis.Number(object.int64_override_type)
        : 0,
      requiredStringViaFieldBehaviorAnnotation: isSet(
        object.required_string_via_field_behavior_annotation,
      )
        ? globalThis.String(
            object.required_string_via_field_behavior_annotation,
          )
        : "",
      outputOnlyStringViaFieldBehaviorAnnotation: isSet(
        object.output_only_string_via_field_behavior_annotation,
      )
        ? globalThis.String(
            object.output_only_string_via_field_behavior_annotation,
          )
        : "",
      optionalStringValue: isSet(object.optional_string_value)
        ? globalThis.String(object.optional_string_value)
        : undefined,
      productId: globalThis.Array.isArray(object?.product_id)
        ? object.product_id.map((e: any) => globalThis.String(e))
        : [],
      optionalStringField: isSet(object.optional_string_field)
        ? globalThis.String(object.optional_string_field)
        : "",
      requiredStringField1: isSet(object.required_string_field_1)
        ? globalThis.String(object.required_string_field_1)
        : "",
      requiredStringField2: isSet(object.required_string_field_2)
        ? globalThis.String(object.required_string_field_2)
        : "",
      requiredFieldBehaviorJsonName: isSet(
        object.required_field_behavior_json_name_custom,
      )
        ? globalThis.String(object.required_field_behavior_json_name_custom)
        : "",
      requiredFieldSchemaJsonName: isSet(
        object.required_field_schema_json_name_custom,
      )
        ? globalThis.String(object.required_field_schema_json_name_custom)
        : "",
      trailingOnly: isSet(object.trailing_only)
        ? globalThis.String(object.trailing_only)
        : "",
      trailingOnlyDot: isSet(object.trailing_only_dot)
        ? globalThis.String(object.trailing_only_dot)
        : "",
      trailingBoth: isSet(object.trailing_both)
        ? globalThis.String(object.trailing_both)
        : "",
      trailingMultiline: isSet(object.trailing_multiline)
        ? globalThis.String(object.trailing_multiline)
        : "",
      uuids: globalThis.Array.isArray(object?.uuids)
        ? object.uuids.map((e: any) => globalThis.String(e))
        : [],
    };
  },
// ...
};

function createBaseABitOfEverything_Nested(): ABitOfEverything_Nested {
  return { name: "", amount: 0, ok: 0 };
}

export const ABitOfEverything_Nested: MessageFns<ABitOfEverything_Nested> = {
  fromJSON(object: any): ABitOfEverything_Nested {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      amount: isSet(object.amount) ? globalThis.Number(object.amount) : 0,
      ok: isSet(object.ok)
        ? aBitOfEverything_Nested_DeepEnumFromJSON(object.ok)
        : 0,
    };
  },

  toJSON(message: ABitOfEverything_Nested): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.amount !== 0) {
      obj.amount = Math.round(message.amount);
    }
    if (message.ok !== 0) {
      obj.ok = aBitOfEverything_Nested_DeepEnumToJSON(message.ok);
    }
    return obj;
  },

  create(base?: DeepPartial<ABitOfEverything_Nested>): ABitOfEverything_Nested {
    return ABitOfEverything_Nested.fromPartial(base ?? {});
  },
  fromPartial(
    object: DeepPartial<ABitOfEverything_Nested>,
  ): ABitOfEverything_Nested {
    const message = createBaseABitOfEverything_Nested();
    message.name = object.name ?? "";
    message.amount = object.amount ?? 0;
    message.ok = object.ok ?? 0;
    return message;
  },
};
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;
// ...
export interface MessageFns<T> {
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}